Moved customer and video selection up to new Store component.
